+++ packages/currency-utils/src/currencyUtils.ts
@@ -1,9 +1,11 @@
 import { Connection, PublicKey, Cluster } from "@solana/web3.js";
 import { getAssociatedTokenAddressSync, getAccount } from "@solana/spl-token";
 import {
-  getPythProgramKeyForCluster,
-  PythHttpClient,
-} from "@pythnetwork/client";
+  PythSolanaReceiverProgram,
+  pythSolanaReceiverIdl,
+} from "@pythnetwork/pyth-solana-receiver";
+import { AnchorProvider, Program, Wallet } from "@coral-xyz/anchor";
+import { token } from "@coral-xyz/anchor/dist/cjs/utils";
 
 export const getBalance = async ({
   pubKey,
@@ -33,15 +35,17 @@ export const getOraclePrice = async ({
   cluster: Cluster;
   connection: Connection;
 }) => {
-  const pythPublicKey = getPythProgramKeyForCluster(cluster);
-  const pythClient = new PythHttpClient(connection, pythPublicKey);
-  const data = await pythClient.getData();
+  const pythProgram: Program<PythSolanaReceiverProgram> = new Program(
+    pythSolanaReceiverIdl,
+    new PublicKey("rec5EKMGg6MxZYaMdyBfgwp4d5rB9T1VQH5pJv5LtFJ"),
+    new AnchorProvider(connection, {} as Wallet, {
+      skipPreflight: true,
+    })
+  );
 
-  let symbol = "";
-  switch (tokenType) {
-    case "HNT":
-      symbol = "Crypto.HNT/USD";
-  }
+  const data = await pythProgram.account.priceUpdateV2.fetch(
+    new PublicKey("DQ4C1tzvu28cwo1roN1Wm6TW35sfJEjLh517k3ZeWevx")
+  );
 
-  return data.productPrice.get(symbol);
+  return data;
 };