+++ packages/helium-entity-manager-sdk/src/resolvers.ts
@@ -35,10 +35,7 @@ export const heliumEntityManagerResolvers = combineResolvers(
   }),
   resolveIndividual(async ({ path, provider }) => {
     if (path[path.length - 1] == "dntPrice") {
-      if (provider.connection.rpcEndpoint.includes("devnet") || provider.connection.rpcEndpoint.includes("test")) {
-        return new PublicKey("BmUdxoioVgoRTontomX8nBjWbnLevtxeuBYaLipP8GTQ");
-      }
-      return new PublicKey("JBaTytFv1CmGNkyNiLu16jFMXNZ49BGfy4bYAYZdkxg5");
+      return new PublicKey("DQ4C1tzvu28cwo1roN1Wm6TW35sfJEjLh517k3ZeWevx");
     }
   }),
   resolveIndividual(async ({ path, args, accounts, provider }) => {