+++ packages/xnft-hotspot/src/constants/index.tsx
@@ -1,2 +1,0 @@
-export { base64LightLogo } from "./base64-light-logo";
-export { base64DarkLogo } from "./base64-dark-logo";